--- a/Jint/Runtime/Interpreter/EvaluationContext.cs
+++ b/Jint/Runtime/Interpreter/EvaluationContext.cs
@@ -44,4 +44,7 @@ internal sealed class EvaluationContext
         Target = null;
         Completion = CompletionType.Normal;
     }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public bool IsAbrupt() => Completion != CompletionType.Normal;
 }