… Jint/Runtime/Interpreter/Expressions/JintBinaryExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintCallExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintConstantExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintFunctionExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintIdentifierExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintImportExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLiteralExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLogicalAndExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLogicalOrExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintMemberExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintMetaPropertyExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintNewExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintObjectExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintSpreadExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintSuperExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintTaggedTemplateExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintTemplateLiteralExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintThisExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintUnaryExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintUpdateExpression.cs
… Jint/Runtime/Interpreter/Expressions/NullishCoalescingExpression.cs
… Jint/Runtime/Interpreter/JintFunctionDefinition.cs
… Jint/Runtime/Interpreter/Statements/JintClassDeclarationStatement.cs
… Jint/Runtime/Interpreter/Statements/JintDoWhileStatement.cs
… Jint/Runtime/Interpreter/Statements/JintExportDefaultDeclaration.cs
… Jint/Runtime/Interpreter/Statements/JintExpressionStatement.cs
… Jint/Runtime/Interpreter/Statements/JintForInForOfStatement.cs
… Jint/Runtime/Interpreter/Statements/JintForStatement.cs
… Jint/Runtime/Interpreter/Statements/JintIfStatement.cs