While it felt like a good idea at the time, it has negative effect on stack handling and also creates some perf hit.

## Jint.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|3d-cube|234.49 ms|0.403 ms|49017.03 KB|
| **New** |	|	| **218.62 ms (-7%)** | **0.296 ms** | **49017.03 KB (0%)** |
| Old |Run|3d-morph|176.66 ms|0.218 ms|48367.5 KB|
| **New** |	|	| **162.56 ms (-8%)** | **0.289 ms** | **48368.18 KB (0%)** |
| Old |Run|3d-raytrace|176.66 ms|0.344 ms|88376.62 KB|
| **New** |	|	| **163.68 ms (-7%)** | **0.324 ms** | **88376.56 KB (0%)** |
| Old |Run|access-binary-trees|78.34 ms|0.181 ms|60933.53 KB|
| **New** |	|	| **77.39 ms (-1%)** | **0.151 ms** | **60932.51 KB (0%)** |
| Old |Run|access-fannkuch|488.22 ms|2.301 ms|136.34 KB|
| **New** |	|	| **453.04 ms (-7%)** | **2.400 ms** | **136.34 KB (0%)** |
| Old |Run|access-nbody|205.19 ms|0.371 ms|63665.94 KB|
| **New** |	|	| **194.53 ms (-5%)** | **0.408 ms** | **63665.8 KB (0%)** |
| Old |Run|access-nsieve|168.77 ms|0.187 ms|21517.44 KB|
| **New** |	|	| **165.07 ms (-2%)** | **0.191 ms** | **21520.51 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|157.58 ms|0.545 ms|56949.04 KB|
| **New** |	|	| **157.10 ms (0%)** | **0.738 ms** | **56949.04 KB (0%)** |
| Old |Run|bitops-bits-in-byte|251.87 ms|0.376 ms|37045.07 KB|
| **New** |	|	| **236.18 ms (-6%)** | **0.676 ms** | **37044.95 KB (0%)** |
| Old |Run|bitops-bitwise-and|139.45 ms|0.484 ms|55938.79 KB|
| **New** |	|	| **152.61 ms (+9%)** | **0.313 ms** | **55938.76 KB (0%)** |
| Old |Run|bitops-nsieve-bits|235.63 ms|0.830 ms|54072.5 KB|
| **New** |	|	| **213.20 ms (-10%)** | **0.915 ms** | **54072.5 KB (0%)** |
| Old |Run|contr(...)rsive [21]|102.28 ms|0.188 ms|83186.02 KB|
| **New** |	|	| **103.48 ms (+1%)** | **0.262 ms** | **83184.26 KB (0%)** |
| Old |Run|crypto-aes|149.30 ms|0.168 ms|14004.18 KB|
| **New** |	|	| **139.87 ms (-6%)** | **0.325 ms** | **14004.03 KB (0%)** |
| Old |Run|crypto-md5|103.88 ms|0.340 ms|77906.85 KB|
| **New** |	|	| **98.70 ms (-5%)** | **0.200 ms** | **77908.26 KB (0%)** |
| Old |Run|crypto-sha1|103.84 ms|0.231 ms|64567.48 KB|
| **New** |	|	| **98.68 ms (-5%)** | **0.170 ms** | **64567.52 KB (0%)** |
| Old |Run|date-format-tofte|108.50 ms|1.043 ms|43346.05 KB|
| **New** |	|	| **102.18 ms (-6%)** | **0.105 ms** | **43346.05 KB (0%)** |
| Old |Run|date-format-xparb|59.25 ms|0.624 ms|24848.68 KB|
| **New** |	|	| **55.39 ms (-7%)** | **0.130 ms** | **24847.7 KB (0%)** |
| Old |Run|math-cordic|361.44 ms|1.372 ms|81978.58 KB|
| **New** |	|	| **336.49 ms (-7%)** | **0.639 ms** | **81976.24 KB (0%)** |
| Old |Run|math-partial-sums|130.78 ms|0.684 ms|50095.98 KB|
| **New** |	|	| **132.01 ms (+1%)** | **0.274 ms** | **50096.08 KB (0%)** |
| Old |Run|math-spectral-norm|129.94 ms|0.240 ms|51833.97 KB|
| **New** |	|	| **133.59 ms (+3%)** | **0.604 ms** | **51833.87 KB (0%)** |
| Old |Run|regexp-dna|106.16 ms|0.472 ms|17667.54 KB|
| **New** |	|	| **105.45 ms (-1%)** | **0.231 ms** | **17668.97 KB (0%)** |
| Old |Run|string-base64|107.50 ms|0.187 ms|11241.92 KB|
| **New** |	|	| **105.69 ms (-2%)** | **0.235 ms** | **11243.11 KB (0%)** |
| Old |Run|string-fasta|203.13 ms|0.326 ms|148053.52 KB|
| **New** |	|	| **194.59 ms (-4%)** | **0.321 ms** | **148053.38 KB (0%)** |
| Old |Run|string-tagcloud|73.63 ms|0.210 ms|44033.8 KB|
| **New** |	|	| **69.67 ms (-5%)** | **0.400 ms** | **44035 KB (0%)** |
| Old |Run|string-unpack-code|71.39 ms|0.146 ms|74420.91 KB|
| **New** |	|	| **70.46 ms (-1%)** | **0.089 ms** | **74422 KB (0%)** |
| Old |Run|strin(...)input [21]|81.44 ms|0.190 ms|39744.77 KB|
| **New** |	|	| **79.62 ms (-2%)** | **0.330 ms** | **39746.35 KB (0%)** |